Refactor - Component imports and seprate Context Provider
