+++ src/Components/index.js
@@ -1,0 +1,7 @@
+export { HomePage } from "./HomePage/HomePage";
+export { Navbar } from "./Navbar/Navbar";
+export { NavPane } from "./NavPane/NavPane";
+export { Playlists } from "./Playlists/Playlists";
+export { VideoCard } from "./VideoCard/VideoCard";
+export { Watch } from "./Watch/Watch";
+export { User } from "./User/User";
