+++ src/styles.css
@@ -18,7 +18,6 @@ body {
 }
 .fa,
 .far {
-  fill: currentColor;
   width: 1.5rem;
   height: 1.5rem;
 }
